+++ source/Collection.ts
@@ -5,12 +5,12 @@ import { Character } from './Character';
  * split a list of Value instances on every token
  *
  * @export
- * @param {Value[]} list
- * @param {...Value[]} tokens
- * @returns {Value[][]}
+ * @param {Array<Value>} list
+ * @param {...Array<Value>} tokens
+ * @returns {Array<Array<Value>>}
  */
-export function split(list: Value[], ...tokens: Value[]): Value[][] {
-	return list.reduce((carry: Value[][], value: Value) => {
+export function split(list: Array<Value>, ...tokens: Array<Value>): Array<Array<Value>> {
+	return list.reduce((carry: Array<Array<Value>>, value: Value) => {
 		const match = tokens.includes(value);
 		const pos = carry.length - Number(!(match || !carry.length));
 
@@ -24,11 +24,11 @@ export function split(list: Value[], ...tokens: Value[]): Value[][] {
  * trim leading and trailing tokens from a list of Value instances
  *
  * @export
- * @param {Value[]} list
- * @param {...Value[]} tokens
- * @returns {Value[]}
+ * @param {Array<Value>} list
+ * @param {...Array<Value>} tokens
+ * @returns {Array<Value>}
  */
-export function trim(list: Value[], ...tokens: Value[]): Value[] {
+export function trim(list: Array<Value>, ...tokens: Array<Value>): Array<Value> {
 	let before: number = 0;
 	let after: number = list.length;
 
@@ -43,16 +43,16 @@ export function trim(list: Value[], ...tokens: Value[]): Value[] {
  *
  * @export
  * @param {TemplateStringsArray} strings
- * @param {...unknown[]} values
- * @returns {Value[]}
+ * @param {...Array<unknown>} values
+ * @returns {Array<Value>}
  */
-export function interleave(strings: TemplateStringsArray, ...values: unknown[]): Value[] {
+export function interleave(strings: TemplateStringsArray, ...values: Array<unknown>): Array<Value> {
 	return strings.reduce(
 		(carry, value, index) =>
 			carry.concat(
 				Array.from(value, (c) => Character.from(c)),
 				index < values.length ? Value.from(values[index]) : []
 			),
-		[] as Value[]
+		[] as Array<Value>
 	);
 }