+++ source/Filters.ts
@@ -1,13 +1,13 @@
-export type FilterFunction = (...args: unknown[]) => boolean;
+export type FilterFunction = (...args: Array<unknown>) => boolean;
 
 /**
  * Determines whether not all values consist of '-' (with a minimum of two)
  *
  * @export
- * @param {...unknown[]} values
+ * @param {...Array<unknown>} values
  * @returns {boolean}
  */
-export function divider(...values: unknown[]): boolean {
+export function divider(...values: Array<unknown>): boolean {
 	return values.every((value) => !/^-{2,}$/.test(String(value)));
 }
 
@@ -15,9 +15,9 @@ export function divider(...values: unknown[]): boolean {
  * Determines whether not all values are undefined
  *
  * @export
- * @param {...unknown[]} values
+ * @param {...Array<unknown>} values
  * @returns {boolean}
  */
-export function defined(...values: unknown[]): boolean {
+export function defined(...values: Array<unknown>): boolean {
 	return values.some((cell) => typeof cell !== 'undefined');
 }