+++ source/index.ts
@@ -1,17 +1,23 @@
 import { interleave } from './Collection';
 import { records } from './Table';
-import { divider, defined } from './Filters';
+import { divider, defined, FilterFunction } from './Filters';
+
+type Item = { [key: string]: unknown };
+type MapperFunction<T = unknown> = (value: T | string | unknown) => T;
+type Mapper<T extends Item> = { [K in keyof T]: MapperFunction<T[K]> };
+type Params = Parameters<typeof interleave>;
+type Table<T extends Item = Item> = { (...args: Params): Array<T> };
 
 /**
- * create a new template literal function parsing tables with a custom set of filters
+ * Create a new template literal function parsing tables with a custom set of filters
  *
  * @export
- * @param {...((...args: unknown[]) => boolean)[]} filters
- * @returns {(...args: Parameters<typeof interleave>) => { [key: string]: unknown }[]}
+ * @param {...Array<Filter>} filters
+ * @returns {(...args: Params) => Array<Item>}
  */
-export function create(...filters: ((...args: unknown[]) => boolean)[]): (...args: Parameters<typeof interleave>) => { [key: string]: unknown }[] {
-	return <T extends { [key: string]: unknown }>(...args: Parameters<typeof interleave>) =>
-		records<T>(interleave(...args), ...filters) as T[];
+export function create(...filters: Array<FilterFunction>): (...args: Params) => Array<Item> {
+	return <T extends Item>(...args: Params) =>
+		records<T>(interleave(...args), ...filters) as Array<T>;
 }
 
 /**
@@ -21,10 +27,10 @@ export function create(...filters: ((...args: unknown[]) => boolean)[]): (...arg
  *
  * @template T
  * @param {TemplateStringsArray} strings
- * @param {...unknown[]} values
- * @returns {T[]}
+ * @param {...Array<unknown>} values
+ * @returns {Array<T>}
  */
-export const empty = create(divider);
+export const empty = <Table>create(divider);
 
 /**
  * Table template literal parser, exludes both any divider row and
@@ -32,8 +38,22 @@ export const empty = create(divider);
  *
  * @template T
  * @param {TemplateStringsArray} strings
- * @param {...unknown[]} values
- * @returns {T[]}
+ * @param {...Array<unknown>} values
+ * @returns {Array<T>}
  */
-export const table = Object.assign(create(divider, defined), { empty });
+export const table = <Table & { empty: Table }>Object.assign(create(divider, defined), { empty });
 export default table;
+
+/**
+ * Create a new table template literaral parser, mapping specified keys
+ *
+ * @export
+ * @template T
+ * @param {Mapper<T>} mapper
+ * @return {Table<T>}
+ */
+export function mapper<T extends Item>(mapper: Partial<Mapper<T>>): Table<T> {
+	const mapping = <MapperFunction<T>>Object.keys(mapper).reduce((carry, key) => (item: T) => carry({ ...item, [key]: (mapper as Mapper<T>)[key](item[key]) }), (item: T): T => item);
+
+	return (...args: Params): Array<T> => table(...args).map(mapping);
+}