+++ CHANGELOG.md
@@ -4,24 +4,55 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.0] - unreleased
+## [unreleased]
+
+### Added
+- Support for return type definition
+- mapper to create a table parser which maps the returned properties
+
+### Changed
+- Updated dependencies
+
+
+## [3.0.1] - 2021-01-01
+
+### Changed
+- Internal type usage
+
+## [3.0.0] - 2020-12-30
+
 ### Changed
 - The package was rewritten to Typescript, this means the import/require mechanics have changed
 
+
 ## [2.0.0] - 2019-01-03
+
 ### Fixed
 - Empty cells are now represented as undefined
 - Cells with mixed content (e.g. string and variable, or multiple variables) are now always represented as string
 
-## [1.0.2 - 1.0.4] - 2018-24
+
+## [1.0.2 - 1.0.4] - 2018-12-24
+
 ### Fixed
 - documentation
 - package.json repository links
 
+
 ## [1.0.1] - 2018-12-24
+
 ### Fixed
 - main script entry was named wrong
 - license was missing
 
+
 ## [1.0.0] - 2018-12-24
 - Initial release
+
+[unreleased]: https://github.com/rspieker/template-literal-table/compare/v3.0.1...HEAD
+[3.0.1]: https://github.com/rspieker/template-literal-table/compare/v3.0.0...v3.0.1
+[3.0.0]: https://github.com/rspieker/template-literal-table/compare/v2.0.0...v3.0.0
+[2.0.0]: https://github.com/rspieker/template-literal-table/compare/v1.0.4...v2.0.0
+[1.0.2 - 1.0.4]: https://github.com/rspieker/template-literal-table/compare/v1.0.1...v1.0.4
+[1.0.1]: https://github.com/rspieker/template-literal-table/compare/v1.0.0...v1.0.1
+[1.0.0]: https://github.com/rspieker/template-literal-table/releases/tag/v1.0.0